Add missing tests for transfers, metafunctions and more
